requirements.txt
numpy==1.26.0
Jinja2==3.1.2
Flask==3.0.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
urllib3==2.0.6
transformers[torch]
faiss-cpu==1.7.4